+++ lib/news_object.dart
@@ -1,28 +1,35 @@
 import 'dart:convert';
 
 import 'enums.dart';
+import 'news_function.dart';
 import 'range_class.dart';
 import 'subject_code.dart';
 
-class NewsGlobal {
+class NewsCore {
   String title;
+  String newsType;
   String contentHtml;
   String content;
-  int date;
+  int datePublished;
+  int dateFetched;
   List<NewsResource> resources = [];
 
-  NewsGlobal.createDefault()
-      : title = '',
+  NewsCore.createDefault()
+      : newsType = 'unknown',
+        title = '',
         contentHtml = '',
         content = '',
-        date = 0,
+        datePublished = 0,
+        dateFetched = 0,
         resources = [];
 
-  NewsGlobal({
+  NewsCore({
+    this.newsType = 'unknown',
     this.title = '',
     this.contentHtml = '',
     this.content = '',
-    this.date = 0,
+    this.datePublished = 0,
+    this.dateFetched = 0,
     required this.resources,
   });
 
@@ -31,11 +38,13 @@ class NewsGlobal {
     var resTemp = resources.reversed.where((p) => p.type == "link").toList();
 
     for (var linkItem in resTemp) {
-      result = result.replaceRange(
-        linkItem.position,
-        linkItem.position + linkItem.text.length,
-        "[${linkItem.text}](${linkItem.content})",
-      );
+      if (linkItem.type == 'link') {
+        result = result.replaceRange(
+          linkItem.position,
+          linkItem.position + linkItem.text.length,
+          "[${linkItem.text}](${linkItem.content})",
+        );
+      }
     }
 
     return result;
@@ -44,28 +53,71 @@ class NewsGlobal {
   Map<String, dynamic> toMap() {
     final result = <String, dynamic>{};
 
+    result.addAll({'news_type': newsType});
     result.addAll({'title': title});
     result.addAll({'content_html': contentHtml});
     result.addAll({'content': content});
-    result.addAll({'date': date});
+    result.addAll({'date_published': datePublished});
+    result.addAll({'date_fetched': dateFetched});
     result.addAll({'resources': resources.map((x) => x.toMap()).toList()});
 
     return result;
   }
 
-  factory NewsGlobal.fromMap(Map<String, dynamic> map) {
-    return NewsGlobal(
+  factory NewsCore.fromMap(Map<String, dynamic> map) {
+    return NewsCore(
+      newsType: map['news_type'] ?? 'unknown',
       title: map['title'] ?? '',
       contentHtml: map['content_html'] ?? '',
       content: map['content'] ?? '',
-      date: map['date']?.toInt() ?? 0,
+      datePublished: map['date_published']?.toInt() ?? 0,
+      dateFetched: map['date_fetched']?.toInt() ?? 0,
       resources: List<NewsResource>.from(map['resources']?.map((x) => NewsResource.fromMap(x))),
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory NewsGlobal.fromJson(String source) => NewsGlobal.fromMap(json.decode(source));
+  factory NewsCore.fromJson(String source) => NewsCore.fromMap(json.decode(source));
+
+  NewsSubject convertToNewsSubject() {
+    try {
+      if (newsType != NewsType.subject.toString()) {
+        return NewsSubject(
+          datePublished: this.datePublished,
+          title: this.title,
+          contentHtml: this.contentHtml,
+          content: this.content,
+          resources: this.resources.toList(),
+          dateFetched: this.dateFetched,
+          // Lecturer name
+          lecturerName: NewsFunction.getNewsLecturerName(this.title),
+          // Lecturer gender
+          lecturerGender: NewsFunction.getNewsLecturerGender(this.title),
+          // Check if is make up or leaving subject lessons.
+          lessonStatus: NewsFunction.getNewsLessonStatus(this.content),
+          // Affected class (got from title).
+          affectedClasses: NewsFunction.getNewsAffectedSubjects(this.title),
+          // Lesson (Works only if lesson status is leaving and make up)
+          affectedLessons: NewsFunction.getNewsAffectedLessons(this.content),
+          // Date (Works only if lesson status is leaving and make up)
+          affectedDate: NewsFunction.getNewsAdjustDate(this.content),
+          // Room (Works only if lesson status is maked up)
+          affectedRoom: NewsFunction.getNewsMakeupRoom(this.content),
+        );
+      } else
+        throw Exception('This news isn\'t for subject news.');
+    } catch (_) {
+      return NewsSubject(
+        datePublished: this.datePublished,
+        title: this.title,
+        contentHtml: this.contentHtml,
+        content: this.content,
+        dateFetched: this.dateFetched,
+        resources: this.resources.toList(),
+      );
+    }
+  }
 }
 
 class NewsResource {
@@ -106,7 +158,7 @@ class NewsResource {
   factory NewsResource.fromJson(String source) => NewsResource.fromMap(json.decode(source));
 }
 
-class NewsSubject extends NewsGlobal {
+class NewsSubject extends NewsCore {
   List<SubjectAffected> affectedClasses;
   int affectedDate;
   LessonStatus lessonStatus;
@@ -116,19 +168,22 @@ class NewsSubject extends NewsGlobal {
   LecturerGender lecturerGender;
 
   NewsSubject({
+    super.newsType = 'unknown',
     super.title = '',
     super.contentHtml = '',
     super.content = '',
-    super.date = 0,
+    super.datePublished = 0,
+    super.dateFetched = 0,
     required super.resources,
-    required this.affectedClasses,
+    List<SubjectAffected>? affectedClasses,
     this.affectedDate = 0,
     this.lessonStatus = LessonStatus.unknown,
-    required this.affectedLessons,
+    RangeInt? affectedLessons,
     this.affectedRoom = '',
     this.lecturerName = '',
     this.lecturerGender = LecturerGender.unknown,
-  });
+  })  : affectedClasses = affectedClasses ?? [],
+        affectedLessons = affectedLessons ?? RangeInt();
 
   NewsSubject.createDefault()
       : affectedClasses = [],
@@ -144,10 +199,12 @@ class NewsSubject extends NewsGlobal {
   Map<String, dynamic> toMap() {
     final result = <String, dynamic>{};
 
+    result.addAll({'news_type': newsType});
     result.addAll({'title': super.title});
     result.addAll({'content_html': super.contentHtml});
     result.addAll({'content': super.content});
-    result.addAll({'date': super.date});
+    result.addAll({'date_published': super.datePublished});
+    result.addAll({'date_fetched': dateFetched});
     result.addAll({'resources': super.resources.map((x) => x.toMap()).toList()});
     result.addAll({'affected_class': affectedClasses.map((x) => x.toMap()).toList()});
     result.addAll({'affected_date': affectedDate});
@@ -162,10 +219,12 @@ class NewsSubject extends NewsGlobal {
 
   factory NewsSubject.fromMap(Map<String, dynamic> map) {
     return NewsSubject(
+      newsType: map['news_type'] ?? 'unknown',
       title: map['title'] ?? '',
       contentHtml: map['content_html'] ?? '',
       content: map['content'] ?? '',
-      date: map['date']?.toInt() ?? 0,
+      datePublished: map['date_published']?.toInt() ?? 0,
+      dateFetched: map['date_fetched']?.toInt() ?? 0,
       resources: List<NewsResource>.from(map['resources']?.map((x) => NewsResource.fromMap(x))),
       affectedClasses: List<SubjectAffected>.from(map['affected_class']?.map((x) => SubjectAffected.fromMap(x))),
       affectedDate: map['affected_date']?.toInt() ?? 0,
@@ -173,7 +232,7 @@ class NewsSubject extends NewsGlobal {
         (element) => element.value == (map['status'] ?? 0),
         orElse: () => LessonStatus.unknown,
       ),
-      affectedLessons: map['affectedLessons'] != null ? RangeInt.fromMap(map['affected_lessons']) : RangeInt(start: 0, end: 0),
+      affectedLessons: map['affected_lessons'] != null ? RangeInt.fromMap(map['affected_lessons']) : RangeInt(start: 0, end: 0),
       affectedRoom: map['makeup_room'] ?? '',
       lecturerName: map['lecturer_name'] ?? '',
       lecturerGender: LecturerGender.values.firstWhere(